+++ Demos/AuTest/Database.cs
@@ -1,7 +1,3 @@
-// ────── ╔╗
-// ╔═╦╦═╦╦╬╣ Database.cs
-// ║║║║╬║╔╣║ <<TODO>>
-// ╚╩═╩═╩╝╚╝ ───────────────────────────────────────────────────────────────────────────────────────
 using Nori;
 namespace Flux;
