+++ Demos/AuTest/Program.cs
@@ -1,7 +1,3 @@
-// ────── ╔╗
-// ╔═╦╦═╦╦╬╣ Program.cs
-// ║║║║╬║╔╣║ <<TODO>>
-// ╚╩═╩═╩╝╚╝ ───────────────────────────────────────────────────────────────────────────────────────
 using System.Reflection;
 using Nori;
 namespace AuTest;
