+++ Test/IO/TCurlIO.cs
@@ -1,3 +1,7 @@
+// ────── ╔╗
+// ╔═╦╦═╦╦╬╣ TCurlIO.cs
+// ║║║║╬║╔╣║ Curl I/O tests
+// ╚╩═╩═╩╝╚╝ ───────────────────────────────────────────────────────────────────────────────────────
 namespace Nori.Testing;
 
 [Fixture (24, "CURL tests for Dwg", "IO")]
