+++ Wad/AuManifest.txt
@@ -1,4 +1,4 @@
-Dwg
+Dwg2
   Layers Styles Blocks -Bound CurrentLayer.Name Ents -FillInterior Grid
 Layer2
   Name Color Linetype Visible
